+++ quacc/recipes/emt/core.py
@@ -6,7 +6,7 @@ NOTE: This set of minimal recipes is mainly for demonstration purposes.
 from __future__ import annotations
 
 import warnings
-from copy import deepcopy
+from typing import Literal
 
 import covalent as ct
 from ase import Atoms
@@ -19,14 +19,16 @@ from quacc.util.calc import run_ase_opt, run_calc
 
 
 @ct.electron
-def static_job(atoms: Atoms, calc_kwargs: dict | None = None) -> dict:
+def static_job(
+    input_atoms: Atoms | dict[Literal["atoms"], Atoms], calc_kwargs: dict | None = None
+) -> dict:
     """
     Carry out a static calculation.
 
     Parameters
     ----------
-    atoms
-        Atoms object
+    input_atoms
+        Atoms object or a dictionary with the key "atoms" and an Atoms object as the value
     calc_kwargs
         Dictionary of custom kwargs for the EMT calculator
 
@@ -35,12 +37,11 @@ def static_job(atoms: Atoms, calc_kwargs: dict | None = None) -> dict:
     dict
         Dictionary of results from `quacc.schemas.ase.summarize_run`
     """
-
+    input_atoms = input_atoms["atoms"] if isinstance(input_atoms, dict) else input_atoms
     calc_kwargs = calc_kwargs or {}
-    input_atoms = deepcopy(atoms)
 
-    atoms.calc = EMT(**calc_kwargs)
-    atoms = run_calc(atoms)
+    input_atoms.calc = EMT(**calc_kwargs)
+    atoms = run_calc(input_atoms)
 
     return summarize_run(
         atoms,
@@ -51,7 +52,7 @@ def static_job(atoms: Atoms, calc_kwargs: dict | None = None) -> dict:
 
 @ct.electron
 def relax_job(
-    atoms: Atoms,
+    input_atoms: Atoms | dict[Literal["atoms"], Atoms],
     relax_cell: bool = True,
     calc_kwargs: dict | None = None,
     opt_swaps: dict | None = None,
@@ -61,8 +62,8 @@ def relax_job(
 
     Parameters
     ----------
-    atoms
-        Atoms object
+    input_atoms
+        Atoms object or a dictionary with the key "atoms" and an Atoms object as the value
     relax_cell
         Whether to relax the cell
     calc_kwargs
@@ -76,24 +77,24 @@ def relax_job(
     dict
         Dictionary of results from quacc.schemas.ase.summarize_opt_run
     """
-
+    input_atoms = input_atoms["atoms"] if isinstance(input_atoms, dict) else input_atoms
     calc_kwargs = calc_kwargs or {}
     opt_swaps = opt_swaps or {}
 
     opt_defaults = {"fmax": 0.01, "max_steps": 1000, "optimizer": FIRE}
     opt_flags = opt_defaults | opt_swaps
 
-    if relax_cell and not atoms.pbc.any():
+    if relax_cell and not input_atoms.pbc.any():
         warnings.warn(
             "Volume relaxation requested but no PBCs found. Ignoring.", UserWarning
         )
         relax_cell = False
 
-    atoms.calc = EMT(**calc_kwargs)
+    input_atoms.calc = EMT(**calc_kwargs)
 
     if relax_cell:
-        atoms = ExpCellFilter(atoms)
+        input_atoms = ExpCellFilter(input_atoms)
 
-    dyn = run_ase_opt(atoms, **opt_flags)
+    dyn = run_ase_opt(input_atoms, **opt_flags)
 
     return summarize_opt_run(dyn, additional_fields={"name": "EMT Relax"})