+++ quacc/recipes/emt/jobflow/slabs.py
@@ -1,6 +1,8 @@
 """Slab recipes for EMT based on Jobflow"""
 from __future__ import annotations
 
+from typing import Literal
+
 from ase import Atoms
 from jobflow import Flow, Response, job
 
@@ -9,7 +11,7 @@ from quacc.util.slabs import make_max_slabs_from_bulk
 
 
 def bulk_to_slabs_flow(
-    atoms: Atoms,
+    input_atoms: Atoms | dict[Literal["atoms"], Atoms],
     slabgen_kwargs: dict | None = None,
     slab_relax_job: job = job(relax_job),
     slab_static_job: job | None = job(static_job),
@@ -27,8 +29,8 @@ def bulk_to_slabs_flow(
 
     Parameters
     ----------
-    atoms
-        Atoms object
+    input_atoms
+        Atoms object or a dictionary with the key "atoms" and an Atoms object as the value
     slabgen_kwargs
         Additional keyword arguments to pass to `make_max_slabs_from_bulk()`
     slab_relax_job
@@ -45,7 +47,7 @@ def bulk_to_slabs_flow(
     Response
         A Response containing Flow of relaxation and static jobs for the generated slabs.
     """
-
+    input_atoms = input_atoms["atoms"] if isinstance(input_atoms, dict) else input_atoms
     slab_relax_kwargs = slab_relax_kwargs or {}
     slab_static_kwargs = slab_static_kwargs or {}
     slabgen_kwargs = slabgen_kwargs or {}
@@ -54,7 +56,7 @@ def bulk_to_slabs_flow(
         slab_relax_kwargs["relax_cell"] = False
 
     # Generate all the slab
-    slabs = make_max_slabs_from_bulk(atoms, **slabgen_kwargs)
+    slabs = make_max_slabs_from_bulk(input_atoms, **slabgen_kwargs)
 
     # Generate the jobs for each slab
     jobs = []
@@ -66,11 +68,11 @@ def bulk_to_slabs_flow(
             outputs.append(job1.output)
         else:
             job1 = slab_relax_job(slab, **slab_relax_kwargs)
-            job2 = slab_static_job(job1.output["atoms"], **slab_static_kwargs)
+            job2 = slab_static_job(job1.output, **slab_static_kwargs)
             jobs += [job1, job2]
             outputs.append(job2.output)
 
     return Response(
-        output={"input_bulk": atoms, "generated_slabs": slabs},
+        output={"input_bulk": input_atoms, "generated_slabs": slabs},
         replace=Flow(jobs, output=outputs),
     )