+++ quacc/recipes/emt/parsl/slabs.py
@@ -1,6 +1,8 @@
 """Slab recipes for EMT"""
 from __future__ import annotations
 
+from typing import Literal
+
 from ase import Atoms
 from parsl import join_app, python_app
 from parsl.app.python import PythonApp
@@ -10,7 +12,7 @@ from quacc.recipes.emt.core import relax_job, static_job
 
 
 def bulk_to_slabs_flow(
-    atoms: Atoms,
+    input_atoms: Atoms | dict[Literal["atoms"], Atoms],
     slabgen_kwargs: dict | None = None,
     slab_relax_app: PythonApp = python_app(relax_job),
     slab_static_app: PythonApp | None = python_app(static_job),
@@ -28,8 +30,8 @@ def bulk_to_slabs_flow(
 
     Parameters
     ----------
-    atoms
-        Atoms object for the structure.
+    input_atoms
+        Atoms object or a dictionary with the key "atoms" and an Atoms object as the value
     slabgen_kwargs
         Additional keyword arguments to pass to make_max_slabs_from_bulk()
     slab_relax_app
@@ -48,6 +50,7 @@ def bulk_to_slabs_flow(
     """
     from quacc.util.slabs import make_max_slabs_from_bulk
 
+    input_atoms = input_atoms["atoms"] if isinstance(input_atoms, dict) else input_atoms
     slab_relax_kwargs = slab_relax_kwargs or {}
     slab_static_kwargs = slab_static_kwargs or {}
     slabgen_kwargs = slabgen_kwargs or {}
@@ -63,13 +66,13 @@ def bulk_to_slabs_flow(
     def _relax_and_static_distributed(slabs):
         return [
             slab_static_app(
-                slab_relax_app(slab, **slab_relax_kwargs)["atoms"],
+                slab_relax_app(slab, **slab_relax_kwargs),
                 **slab_static_kwargs,
             )
             for slab in slabs
         ]
 
-    slabs = make_max_slabs_from_bulk(atoms, **slabgen_kwargs)
+    slabs = make_max_slabs_from_bulk(input_atoms, **slabgen_kwargs)
 
     if slab_static_app is None:
         return _relax_distributed(slabs)