+++ quacc/recipes/emt/prefect/slabs.py
@@ -1,6 +1,9 @@
 """Slab recipes for EMT"""
 from __future__ import annotations
 
+from typing import Literal
+
+from ase import Atoms
 from prefect import flow
 
 from quacc.util.slabs import make_max_slabs_from_bulk
@@ -11,7 +14,7 @@ from quacc.util.slabs import make_max_slabs_from_bulk
 # like in other workflow engines. Relies on #10135 in Prefect
 @flow
 def bulk_to_slabs_flow(
-    atoms,
+    input_atoms: Atoms | dict[Literal["atoms"], Atoms],
     slab_relax_task,
     slab_static_task,
     slabgen_kwargs: dict | None = None,
@@ -29,8 +32,8 @@ def bulk_to_slabs_flow(
 
     Parameters
     ----------
-    atoms
-        Atoms object for the structure.
+    input_atoms
+        Atoms object or a dictionary with the key "atoms" and an Atoms object as the value.
     slab_relax_task
         Default Task to use for the relaxation of the slab structures.
     slab_static_task
@@ -48,6 +51,7 @@ def bulk_to_slabs_flow(
         List of dictionary of results from quacc.schemas.ase.summarize_run
         or quacc.schemas.ase.summarize_opt_run
     """
+    input_atoms = input_atoms["atoms"] if isinstance(input_atoms, dict) else input_atoms
     slab_relax_kwargs = slab_relax_kwargs or {}
     slab_static_kwargs = slab_static_kwargs or {}
     slabgen_kwargs = slabgen_kwargs or {}
@@ -67,7 +71,7 @@ def bulk_to_slabs_flow(
             for slab in slabs
         ]
 
-    slabs = make_max_slabs_from_bulk(atoms, **slabgen_kwargs)
+    slabs = make_max_slabs_from_bulk(input_atoms, **slabgen_kwargs)
 
     if slab_static_task is None:
         return _relax_distributed(slabs)