+++ quacc/recipes/emt/slabs.py
@@ -1,6 +1,8 @@
 """Slab recipes for EMT"""
 from __future__ import annotations
 
+from typing import Literal
+
 import covalent as ct
 from ase import Atoms
 
@@ -9,7 +11,7 @@ from quacc.util.slabs import make_max_slabs_from_bulk
 
 
 def bulk_to_slabs_flow(
-    atoms: Atoms,
+    input_atoms: Atoms | dict[Literal["atoms"], Atoms],
     slabgen_kwargs: dict | None = None,
     slab_relax_electron: ct.electron = relax_job,
     slab_static_electron: ct.electron | None = static_job,
@@ -27,8 +29,8 @@ def bulk_to_slabs_flow(
 
     Parameters
     ----------
-    atoms
-        Atoms object for the structure.
+    input_atoms
+        Atoms object or a dictionary with the key "atoms" and an Atoms object as the value
     slabgen_kwargs
         Additional keyword arguments to pass to make_max_slabs_from_bulk()
     slab_relax_electron
@@ -45,6 +47,7 @@ def bulk_to_slabs_flow(
     list[dict]
         List of dictionary of results from quacc.schemas.ase.summarize_run or quacc.schemas.ase.summarize_opt_run
     """
+    input_atoms = input_atoms["atoms"] if isinstance(input_atoms, dict) else input_atoms
     slab_relax_kwargs = slab_relax_kwargs or {}
     slab_static_kwargs = slab_static_kwargs or {}
     slabgen_kwargs = slabgen_kwargs or {}
@@ -62,13 +65,13 @@ def bulk_to_slabs_flow(
     def _relax_and_static_distributed(slabs):
         return [
             slab_static_electron(
-                slab_relax_electron(slab, **slab_relax_kwargs)["atoms"],
+                slab_relax_electron(slab, **slab_relax_kwargs),
                 **slab_static_kwargs,
             )
             for slab in slabs
         ]
 
-    slabs = ct.electron(make_max_slabs_from_bulk)(atoms, **slabgen_kwargs)
+    slabs = ct.electron(make_max_slabs_from_bulk)(input_atoms, **slabgen_kwargs)
 
     if slab_static_electron is None:
         return _relax_distributed(slabs)