+++ tests/parsl_tests/test_parsl.py
@@ -53,7 +53,7 @@ def test_tutorial1():
     future1 = relax_app(atoms)
 
     # Call App 2, which takes the output of App 1 as input
-    future2 = static_app(future1.result()["atoms"])
+    future2 = static_app(future1)
     future2.result()
     assert future2.done()
 
@@ -100,7 +100,7 @@ def test_tutorial3():
 
     # Define the workflow
     future1 = relax_app(atoms)
-    future2 = bulk_to_slabs_app(future1.result()["atoms"])
+    future2 = bulk_to_slabs_app(future1)
 
     # Print the results
     future2.result()
@@ -123,7 +123,7 @@ def test_tutorial4():
 
     # Define the workflow
     future1 = relax_app(atoms)
-    future2 = bulk_to_slabs_flow(future1.result()["atoms"], slab_static_app=None)
+    future2 = bulk_to_slabs_flow(future1, slab_static_app=None)
 
     # Print the results
     print(future2.result())