+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Changed
 
+- All recipes now support the `Atoms` object being passed in as a `dict[Literal["atoms"], Atoms]`.
 - The `slab_relax_job` kwarg in `recipes.emt.slabs` and `recipes.vasp.slabs` workflows can no longer be set to `None`, as there are few situations where this would be desired.
 - Class-based recipes have been converted to functions since they don't save state or have inheritance.